+++ sdk/metric/internal/aggregate/lastvalue.go
@@ -52,21 +52,52 @@ func (s *lastValueMap[N]) measure(
 	}
 }
 
+type lazyLastValueMap[N int64 | float64] struct {
+	values lazyLimitedSyncMap[*lastValuePoint[N]]
+}
+
+func (s *lazyLastValueMap[N]) measure(
+	ctx context.Context,
+	value N,
+	fltrAttr attribute.Set,
+	droppedAttr []attribute.KeyValue,
+) {
+	lv := s.values.LoadOrStoreAttr(fltrAttr)
+
+	lv.value.Store(value)
+	if !lv.dropExemplars {
+		lv.res.Offer(ctx, value, droppedAttr)
+	}
+}
+
 func newDeltaLastValue[N int64 | float64](
 	limit int,
 	r func(attribute.Set) FilteredExemplarReservoir[N],
 ) *deltaLastValue[N] {
+	newVal := func(attr attribute.Set) *lastValuePoint[N] {
+		res := r(attr)
+		_, isDrop := res.(*dropRes[N])
+		p := &lastValuePoint[N]{
+			res:           res,
+			attrs:         attr,
+			startTime:     now(),
+			dropExemplars: isDrop,
+		}
+		return p
+	}
+	resetFunc := func(v *lastValuePoint[N]) {
+		v.value.Store(0)
+	}
+
 	return &deltaLastValue[N]{
 		newRes: r,
 		start:  now(),
-		hotColdValMap: [2]lastValueMap[N]{
+		hotColdValMap: [2]lazyLastValueMap[N]{
 			{
-				values: limitedSyncMap{aggLimit: limit},
-				newRes: r,
+				values: newLazyLimitedSyncMap[*lastValuePoint[N]](limit, newVal, resetFunc),
 			},
 			{
-				values: limitedSyncMap{aggLimit: limit},
-				newRes: r,
+				values: newLazyLimitedSyncMap[*lastValuePoint[N]](limit, newVal, resetFunc),
 			},
 		},
 	}
@@ -78,7 +109,7 @@ type deltaLastValue[N int64 | float64] struct {
 	start  time.Time
 
 	hcwg          hotColdWaitGroup
-	hotColdValMap [2]lastValueMap[N]
+	hotColdValMap [2]lazyLastValueMap[N]
 }
 
 func (s *deltaLastValue[N]) measure(
@@ -125,7 +156,7 @@ func (s *deltaLastValue[N]) copyAndClearDpts(
 		dPts[i].StartTime = s.start
 		dPts[i].Time = t
 		dPts[i].Value = v.value.Load()
-		collectExemplars[N](&dPts[i].Exemplars, v.res.Collect)
+		collectExemplarsAfter[N](&dPts[i].Exemplars, s.start, v.res.Collect)
 		i++
 		return true
 	})